Fix:アウトプット一覧部分をShowOutputs.vueを作成し、component化
